Initialize module for basic projectComponent class
Currently, only the constructor and the print() methods are defined.

The properties, private and specific to each project, are an ID, a
title, a description, a date of creation (determined automatically),
and a date of last edit (which is initialized with the date of
creation).

Note that the ID is also set automatically, based on the nextId static
property of the class, which is incremented after that, ensuring that
each project is given an unique ID.