Initialize module for basic todoComponent class
Currently, only the constructor and the print() methods are
defined. The setter/getter methods are not implemented yet.

The properties, private and specific to each todo, are an ID, a
title, a description, a date of creation (determined automatically),
a date of last edit (which is initialized with the date of creation),
a due date (optional, initialized to null), a priority (none, low,
medium, high), a state ('todo', 'wip', 'done'), an imminence ('none',
'upcoming', 'due soon', 'today', 'expired'), an associated project
id (initialized to null), and an array of tags.

Note that the ID is also set automatically, based on the nextId static
property of the class, which is incremented after that, ensuring that
each project is given an unique ID.

Also, priority, state, and imminence are specified as an index of
three array static arrays of the class, respectively, which contains
the corresponding possible labels.