Add methods related to tags list in projectComponent class
The property #data.tags is now a Set, as it is meant to contain
only unique data.

Three methods are added to the projectComponent class. hasTag(tag),
addTag(tag), and removeTag(tag).

addTag and removeTag return true if the object is modified, false
otherwise.

Note that to pass a tags list to the constructor argument, the
data.tags must be a Set (although this is not checked yet).