Refactor projectComponent class to inherit from some baseComponent class
Since most of the code/properties/methods of projectComponent class
are shared with todoComponent class (e.g., id, title, description,
tags, ...), create a baseComponent class which defines the common
parts.

Then, projectComponent class inherits from this baseComponent class
and adds the properties specific to it, ie, the list of todos and
the corresponding methods.

For todoComponent, a similar operation will be done.