Add getter methods to todoComponent class
Note that imminence and associatedProjectId have to be revised.

Probably, it's better not to put them in .data, also in view of the
fact that .data is going to be saved in a .json file in a future
release.

Indeed, imminence is a dynamic property that depends only on the
dueDate and the date of today. An user cannot set it.

associatedProjectId should be modified to a reference to the
associated parent. While you cannot save a reference to json, you
can rebuild the relation when loading the data, since the information
project -> todo can be saved. Saving a reference, let you easily
retrieve all the parent info.