Add .parent and .type to baseClass + refactor print method
A .parent property has been added to baseClass (and, so, to
projectClass and todoClass, too), which is initialized to 0.
If a parent reference is specified in the constructor, it is
assigned to it instead.

One may want, eg, to create sub-projects in each project, although
this is not yet implemented. Then, this .parent enables this.

For instance, you can get a chain like:

	null <- Project <- (sub) Project <- Todo <- (subtask) Todo

And you can traverse this chain back to null to:
1) determine the path of the last element in the chian (eg, here,
   the subtask).
2) being able to traverse the chain back to update the last edited
   property of each element when you modify a component.

Note: as of now, Projects cannot have a parent.

A .type property has been added to baseClass. This is used to get
the specific kind of class. It is reassigned when initializing a
projectClass and todoClass.
	'B' <-> baseClass
	'P' <-> projectClass
	'T' <-> todoClass

Finally, the print method has been refactored. Common properties
are now printed based on .printBaseInfo() from baseComponent class.