Handle imminence property of todoComponent class
Instead of being a property of .data property of the todoComponent
object, now there is a private variable .#imminenceIdx (which is
the index of the correpsonding label of todoComponent.imminenceLabels
array).

You can retrieve the label of the index of the current imminence,
but not modify it. You can only update it, based on the number of
days between the due date and now, calling the updateImminence()
method.

Note that updateImminence() must be called every time the due date
is modified, and at midnight, too (todo).