Add domStructure, headerDomComponent, and navDomComponent modules
These modules are used to dynamically create the webpage.

domStructure defines a class to initialize the basic structure,
ie, header, main and footer elements. Currently, only the header
is created.

headerDomComponent defines a class to initialize the header. This
consists of a h1 element with the name of the app, and a nav element
to navigate it.

navDomComponent create the actual navigation menu. For each project
in root.customProjects and for inbox, a button is created to trigger
the rendering of that view (this functionality is yet to be coded).
There are two sections: a predefined div (which currently contains
only the inbox project), and a projects div, which shows the current
projects. This nav will be improved over time.

Note that no style is defined yet.