Add mainDomComponent and projectDomComponent modules
These modules are used to dynamically create the webpage.

mainDomComponent defines a class to initialize the main element.
This will contain some custom views. As of now, it will show some
project details when the user navigates to it. If mainDomObj is an
instance of this class, the actual dom element is mainDomObj.main.

Note that domStructure module has been updated to create the main,
too. In doing that, a reference to the mainDomObject is saved in the
body as:

   document.body.mainDomObj = mainDomObj;

to make this object easily accessible each time there is the need to
modify it.

projectDomComponent creates a div showing some project info which can
be rendered in the main element. As of now, a simple text is shown with
the project info returned by project.print(). A more suitable dom
structure will be created next.

The renderProject() method of mainDomComponent class initialize a
projectDomComponent instance and renders it.
Then, #btnRenderProjectCallback() callback in navDomComponent class can
call:

   document.body.mainDomObj.renderProject(e.target.associatedProject);

to render the required project view.

Note that no style is defined yet.