Add baseDomComponent module
The baseDomComponent is meant to create a dom representation of the baseComponent objects. As of now, only the footer with the date of creation and possible date of last edit are created and put in a footer inside the div representing the object.
There are methods to create the single sub-components (eg, representing the date of creation, etc), or blocks (eg, the above-mentioned footer), or the whole component (with the .init() method, which is called in the constructor.) This way, when a class inherits from this baseDomComponent, it can render part of such elements as needed: it can also redefine the init() method to have a custom structure.

In particular, projectDomComponent class extends baseDomComponent class, and initializes a footer using the initFooter() method inherited from baseDomComponent.

The baseComponent has been added a hasBeenEdited() method to check whether such component has been edited. Other minor adjustments have been made. Note that now a component which has not been edited has null as data.dateOfEdit.