Add dom header initializaton for baseDomComponent
The baseDomComponent does now render also the header with the path
(of the parent elements up to null, to be tested further), the
title, the description, and the tag list.
The corresponding single components are also initializated in
separated methods.

projectDomComponent class exteinherits such methods, too, to
initialize the header in the dom.

Note that the tags and the elements in the path sould be buttons
to render the corresponding content. As these functionality is not
implemented yet, they are simple li elements of an ul list as of
now. This might be improved in the future.

Note that no style is defined yet.