Add status button in the todoDomComponent
The status button is meant to summarize the state (todo, wip,
done) and the priority (none, low, medium, high) of the todo.

The icon shows the state of the todo, the color shows the priority

Note that the button is not functional yet. It will be used to
change the state of the todo.