Use this.constructor.defaultData instead of baseComponent.defaultData in baseComponent class
This way, when using that variable, it refers to the static
defaultData variable defined in the specific class, and not to the
variable defined in baseComponent class only.

Then, in order to make this work, the projectComponent class has
been modified as follows:

1) 'import' the defaultData from baseComponent class when defining
   it in the projectComponent class, ie:

	  static defaultData = {
	    ...baseComponent.defaultData,
	    dueDate: null,
	    priority: 0 /* index of todoComponent.prioritiesLabels array*/,
	    state: 0 /* index of todoComponent.stateLabels array*/,
	  };

2) removing the following line, since it is not needed anymore,
   because of the previous point:

    	this.data = Object.assign({}, projectComponent.defaultData, this.data);

Similar changes have been done in todoComponent class.