Refactor static variables in baseDomComponent, projectDomComponent and todoDomComponent classes
The blockName and cssClasses in there are no more global variables
of the module.

In particular, blockName is now a static variable defined in
each class. You can refer to the one specific to each class using
this.constructor.blockName. Note that you do not need to call the
constructor with blockName anymore.

Moreover, instead of using cssClass.XXX or cssClass.XXX() to
retrieve the css class to use for some label XXX, a specific method
getCssClass(element) is defined in the baseClass, where element is
the label XXX. This method retrieves the specific cssClass based
on this label XXX. Note that in projectDomComponent class and
todoDomComponent class we inherit the cssClass from baseDomComponent
using:

  static cssClass = {
    ...baseDomComponent.cssClass,
    /* ... */
  };

This change is because in the future these classes might be
inherited by other classes, too.

The other global variables in this modules have been moved to
static variables specific to each class, for consistency.

The rest of the code has been updated to reflect the
above-mentioned changes.