Add functionality to the todoDomComponent status button
Clicking the button toggle the state of the todoComponent object,
eg, todo -> work in progress -> done -> todo -> ...

To update the interface, the PubSubJS library has been used. Any
dom element that has been defined based on the todo.state property,
is now subscribed to the 'TODO STATE CHANGE'. When this is
published, the interface is properly adapted. It is the setter
method of todoDomComponent class that actually publish this.