Add baseListComponent, projectListComponent and todoListComponent modules
The baseListComponent defines a class to handle a list of
baseComponent instances. It contains the actual list, a reference
to a parent element, and a title. It includes methods to add/
remove baseComponent elements to/from the list. Note that when
this happens, if the parent is not null, its updateDateOfEdit()
method is called.

The projectListComponent and todoListComponent classes inherit from
baseListComponent class. They only redefine the initItem(data)
method, to add a projectComponent and todoComponent item to the
list, respectively.

The reason of the definition of the above classes to to avoid reuse
the same code, since, in fact, the methods to modify a list of
projects and todos are quite the same.