Add baseListDomComponent and todoListDomComponent modules
The baseListDomComponent defines a class to handle the display of
a baseListComponent instance in the dom, which is to be passed to
the constructor. It includes methods to add items to the displayed
list. However, a method to remove items from the displayed list is
not implemented yet.

The todoListDomComponent class inherit from baseListDomComponent
class. It only redefines the initItemDom(item) method to add a todoDomMiniComponent item to the displayed list, and the
btnRenderItemCallback(e) callback to display the full
todoDomComponent element on click.

Similarly, a projectListDomComponent class will be defined in a
future commit.

The reason of the definition of the above classes to to avoid reuse
the same code.