Add projectDomMiniComponent and projectDomNavComponent modules
These defines the corresponding classes in a similar way of
todoDomMiniComponent class. Instances of this classes will be
used to only show the specific minimal info for a project, when
building the UI.

Note that projectDomNavComponent instances are even more minimal
than projectDomMiniComponent instances, showing only the title of
the project.