Add 'Add new item' button to baseListDomComponent dom element
The button will be used to add a new item to the baseListComponent
object linked to the baseListDomComponent instance. Data will be
obtained from a form that will be shown. All these functionalities
are not implemented yet.

The button is present also in the projectListDomComponent class and
todoListDomComponent class, which inherit from baseListDomComponent
class, altough a custom callback will be used.