Initialize baseFormDomComponent module
This defines a class which creates a dialog to be used to get data
to add a new project or todo.

As of now, only some basic elements are created (eg, title, or
submit button): the actual input elements are not implemented yet.

An instance of this class is created in domStructure.js for testing
purposes. This will be used to show/hide the dialog properly.
In particular, in baseDomComponent class, the addNewItem() method
has been modified to show this dialog as a modal.

Note that as of now, this baseFormDomComponent is used, just to
test the class. In the final version, some projectFormDomComponent
and todoFormDomComponent classes will be used, which are specific
to projects or todos instances.