Add Tag input elements in dialog Elements
A .*_new-tag-btn enables the insertion of a new tag. In particular,
a div element containing an input element for the user to fill the
actual tag and a cancel tag is created.

Note that as of now is not checked whether tags are valid (eg,
they are not repeated). This will be added later.