Add new items on the associated parent object on form submit in dialog elements
A .setObjectData(object, edit = false) method is added to the
baseFormDomComponent class. This method is in charge of setting
the object where the form data will be used. In particular, if
'edit' is false, it means that a new object is to be added to the
parent 'object'. Otherwise, 'object' is to be edited (although
this is not implemented, yet). Due to inheritance, this is true
also for the projectFormDomComponent and todoFormDomComponent
classes, as well.

The baseListDomComponent instances (and the projectListDomComponent
and todoListDomComponent, as well) call the .setObjectData() method
on the associated dialog before opening it to add a new item, to
set the object where the new item has to be added.

The form 'submit' event listener has been updated to add the new
item using the .addItem(data) method that must be present in the
object 'object' mentioned above. Indeed, baseListDomComponent has
already this method.

Note that currently submitting data from the form dialog modifies
the assoicated baseListComponent/projectListDomComponent/
todoListDomComponent only. Then, the current view is not edited,
 yet. You can see the newly created item when that view is
 re-rendered.