Link .addItem() method of baseListComponent class and .addDomItem() of baseListDomComponent class via PubSub
When the addItem(data) method of baseListComponent (or
projectListComponent, or todoListComponent) class is called, a
proper PubSub token is published, which identifies the parent
object of the corresponding list, the type of list and the
performed action.

The baseListDomComponent (or projectListDomComponent, or
todoListDomComponent) instances are subscribed on similar tokens.
When the token is mathced (ie, when a new item is added to the
list corresponding list), their .addDomItem() method is called to
display the new item.