Add priority selection in new todo dialog element
A <select> element is used to select the priority of the new todo.
The possible values are taken directly from the static property
todoComponent.priorityLabels.

The modification is done on the todoFormDomComponent class.

Note that the getDataToSubmit() and resetForm() methods do not
take this modification into account, yet. So no data is used to
create a new todo, nor this select field is reset when the dialog
is closed.

No style is applied, yet.