Add due date selection in new todo dialog element
A <select> element is used to select whether the user wants to
set a due date or not. In the latter case, a input with type=
'date' is shown by which the user can select the desired due
date. Some default dates are shown using a datalist element
(today, tomorrow, next week, next month, next year).

The modification is done on the todoFormDomComponent class.

Note that the getDataToSubmit() and resetForm() methods do not
take this modification into account, yet. So no data is used to
create a new todo, nor this select field is reset when the dialog
is closed.

No style is applied, yet.