Update the resetForm() method of todoFormDomComponent class
A resetFormOtherInfo() method has been defined to reset the dueDate
and priority input in the todoFormDomComponent class.

The resetForm() of the baseFormDomComponent class (from which
todoFormDomComponent class inherits) has been assigned to the
this.resetFormBase() method in the constructor, and then the
this.resetForm() method has been updated, using an arrow function,
calling both this.resetFormBase() and this.resetFormOtherInfo().

Now all the data are reset when the form dialog is closed.