Update the getDataToSubmit() method of todoFormDomComponent class
A getDataToSubmitOtherInfo() method has been defined to get the
data from the dueDate and priority inputs in the
todoFormDomComponent class, to create a new item.

The getDataToSubmit() of the baseFormDomComponent class (from
which todoFormDomComponent class inherit) has been assigned to the
this.getDataToSubmitBase() method in the constructor, and then the
this.getDataToSubmit() method has been updated, using an arrow
function, calling both this.getDataToSubmitBase() and
this.getDataToSubmitOtherInfo().

Now all the required input data are used to create a new todo.