Add list property to baseDomComponent class
list is a reference to the containing list, if present. Otherwise
it is null. Eg, the 'Incoming' project is not a part of any list.
The other projects, instead, are part of a projectListComponent
instance, and the todos of a todoListComponent instance.

This property will simplify the handling of deleting a project/
todo from a list.

Other classes using this class have been modified accordingly.