Fix removeDomItem(item) method of baseListDomComponent and use it to delete items from dom list
The projectListDomComponent and todoListDomComponent classes
inherit from baseListDomComponent class, hence inherit the
removeDomItem() method.

These are composed of projectDomMiniComponent and
todoDomMiniComponent items. When one of such items is set to be
deleted from the projectDomMiniComponent and todoDomMiniComponent
view, use removeDomItem() appropriately to remove the item and
avoiding rendering the whole view. This is handled via PubSub.