Open the correct form dialog when clicking on edit button on baseDomComponent instances
When clicking on the edit button on baseDomComponent instances,
the correct dialog to edit the corresponding object is shown.

Note that the dialog is not filled with the data of the object to
edit, yet, nor it modifies it.

This applies also for the inherited classes (eg, todoDomComponent,
projectDomComponent, todoDomMiniComponent, ...).