Fix minimum input dueDate in todoFormDomComponent instances
So far, the minimum date that could be set on the dueDate input in
todoFormDomComponent instances was the 'today' date. When editing
a todo, a dueDate in the past would be invalid (eg, if the todo is
done or expired). Then, when the form dialog is set in 'edit' mode,
set the minimum date constraint on the dueDate as the date of its
creation. Else, set it to 'today'.