Refactor tag handling in baseFormDomComponent class
Now the tag currently displayed in the form are stored in a Map
object whose keys are the tags themselves (including the empty
string ''), and the values arrays of the corresponding input
elements.

Having a '' keys means that there is an invalid empty tag that the
user must fill. Having an array with more than one item as value
in the map, means that the tag is repeated and the user must remove
one of them.

These conditions are notified to the user using the
.setCustomValidity() and .reportValidity() methods of the input
dom elements.