Add noteComponent module
This module defines a class which istantiate an object that
describes a note.

It is assumed that a note can have only a title, a description,
and some tags.

This class inherits from baseComponent class. It is basically the
same, except for having its own nextId and having a custom type
'N'.

Note: more components will be added next, related to notes. To
test them, some temporary code is added to the homeDomComponent,
to show these notes. It will be removed later on.