Add noteDomComponent module
This module defines a class which instantiate an object that
creates and handle a div to render a note.

This class inherits from baseDomComponent class. It is basically
the same, except for having its own blockName and
associatedDialog method.

Note that the document.body.noteFormDialog used in the
associatedDialog method is currently undefined, but will be
implemented later on.

No style is defined, yet.

A temporary div of a noteDomComponent instance is added to the dom
in the homeDomComponent module, for testing.