Add noteDomMiniComponent module
This module defines a class which instantiates an object that
creates and handles a div to render a mini-view of a note.

This class inherits from noteDomComponent class. Also, it is is basically
the same as projectDomMiniComponent class.

A renderNote(noteObj) method has been added to the mainDomComponent
class. This is used in the  PubSub subscription callback for the
"RENDER NOTE" token, to render a noteDomComponent div. Such token is
emitted in the projectDomMiniComponent class.

No style is defined, yet.

A temporary div of a noteDomMiniComponent instance is added to the dom
in the homeDomComponent module, for testing.