Add noteFormDomComponent module
This module defines a class which instantiates an object that
creates and handles a form to create / edit a note.

This class inherits from baseFormDomComponent class. It is
basically the same, except for having its own blockName and type.

A new noteFormDomComponent is instantiated in domStructure class
(with its form added to the dom) and its reference saved in
document.body.noteFormDialog. This is used in the associatedDialog
method in noteDomComponent class.

No style is defined, yet.