Add noteListDomComponent module
This module defines a class which instantiate an object that
creates and handles a div to render a list of a notes.

This class inherits from baseListDomComponent class. It is
basically the same, except for redefining the initItemDom()
method, to add a noteDomMiniComponent object to the list.

No style is defined, yet.

A temporary noteListComponent() object and the corresponding
noteListDomComponent div is generated and added to the dom
in the homeDomComponent module, for testing. Please note that
instantiating the noteListComponent object here this way reset
each time the home view is rendered. In practice the list won't be
instantiated here.