Add listInComponentMixin module to add object list managements via composition
This module defines a mixin which is used to add the support for
todoList (T), projectList (P), noteList (N) to a given class,
using composition.

In particular, after having defined a class className, you should
call:
   listInComponentMixin(className, whichListArray)
where whichListArray is an array of the codes of the lists you
need (eg, whichListArray = ["T", "P", "N"]).

The following methods are added:

- initAllLists() ---> this must be called in the constructor to
  add the lists of type T, P, and/or N to the data.lists object
  property of the class.
- print() ---> this redefines the print() method to print also the
  new lists

and, based on whichListArray:

- getProjectList(), getTodoList(), getNoteList()
- addToProjectList(data), addToTodoList(data),
  addToNoteList(data)
- removeFromProjectList(obj), removeFromTodoList(obj),
  removeFromNoteList(obj)

The code to handle the todoList in projectComponent class has been
removed. Instead, the methods are now added to that class via
composition using listInComponentMixin function, and the list
object calling initAllLists().

Note that the names of some of these methods have changed. The
very few instances in which they had been used have been fixed now
using the new names.