Refactor the redefinition of methods in inherited classes
Let method() be a method of some parentClass which is inherited by
some childClass. When such method needs to be redefined in the
child class, now it is done in the following way:

  method(data) {
    super.method();
    // add custom code here
  }

where super.method() is the method of the parentClass.

This also fixes a minor bug intoduced probably in one of the last
commits, by which the title of completed todos was not propely
styled.