Fix circular dependencies introduced by listInComponentMixin using the 'internal module pattern'
The introduction of the listInComponentMixin module introduces
some circular import dependencies that might be problematic.

Eg, when you want to create a project with a list of projects,
you get this circular module dependency:

   projectComponent -> listInComponentMixin -> projectListComponent -> projectComponent

For testing purposes, a projectListComponent, todoListComponent
and nodeListComponent have been added to the todoComponent and
nodeComponent. Note that nothing is displayed in the dom yet.
Also, in the final version, you would not actually want a
project list inside a note, so this has to be rearranged. In the
current configuration, though, the original page could not be
rendered at all.

The possible bad effects of these are due to the module loading
order and can be removed applying the so called 'internal module
pattern', which gives full control over the module loading order.

This has been used only for the subset of modules whose import
would introduce circular dependencies, ie, projectComponent,
todoComponent, noteComponent, projectListComponent,
todoListComponent, noteListComponent, and listInComponentMixin.

In particular, they all are imported in the right order in the
fixCircularDependenciesInComponents.js module, as well as
exported. Whenever one of those modules is needed, it has to be
imported directly from this file.

Despite circular dependencies remains, they are only in this new
module, and are resolved natively by Javascript.

see:  https://medium.com/visual-development/how-to-fix-nasty-circular-dependency-issues-once-and-for-all-in-javascript-typescript-a04c987cf0de

A visualization of the dependencies has been obtained using Madge,
and can be visualized in graph.svg.
see: https://github.com/pahen/madge