Add listInDomComponentMixin module to add object list dom management via composition
This module defines a mixin which is used to add the support for
the lists defined in this.obj.list in the instances of a given
class, using composition: a todoDomList, projectDomList, and/or
noteDomList instance is added based on the label of the list.

In particular, after having defined a class className, you
should call:
   listInDomComponentMixin(className);

The following method is added:
- initAllLists() ---> this must be called in the constructor

The code to handle the todoList in projectDomComponent class has
been removed. Instead, the code is now added to that class via
composition using listInComponentMixin function. The same has
been done in todoDomComponent and noteDomComponent.

Circular import dependencies are introduced by
listInDomComponentMixin. For instance, we have:

   projectDomComponent -> listInDomComponentMixin -> projectListDomComponent -> projectDomMiniComponent -> projectDomComponent

They are fixed using the 'internal module pattern' (see the
previous commit), considering only the involved modules:
projectDomComponent, todoDomComponent, noteDomComponent,
projectDomMiniComponent, todoDomMiniComponent,
noteDomMiniComponent, projectListDomComponent,
todoListDomComponent, noteListDomComponent,
listInDomComponentMixin.