Restore the static property nextId of projectComponent, todoComponent and noteComponent classes
This is done when creating a new instance of rootComponent (which
creates the whole logic structure of the page) with a dataJSON
parameter in the constructor.

The dataJSON must have the structure described in the previous
commit, as the one created from the rootComponent.prototype.toJSON()
method. Eg,

	{"nextId":{"projectComponent":4,"todoComponent":5,"noteComponent":0}, ...}

(although not all fields are mandatory).

If no dataJSON parameter is provided, or some properties are
missing, the nextId is initialized to the default (0).