Add itemDataLists parameter to initAllLists method defined by listInComponentMixin to restore its lists content
initAllLists is an object of input parameters for
projectListComponent, todoListComponent, and noteListComponent
class constructors, each one labeled by projectList, todoList, and
noteList, respectively.

Then, the projectComponent, todoComponent, and noteComponent, have
been modified to pass the data.lists argument to the initAllLists
method call in their constructor.

The rootComponent class is now able to restore the whole Inbox and
projects the Project lists.