Add size method to baseListComponent and inherited classes
A .size getter method has been added to the baseListComponent and
the inherited classes, to return the number of items in the
corresponding list.

Some temporary test code has been added to initWebpage.js, to
show the size of the lists (at top level) of the custom
projects. This will be removed later on.