Add match method to baseComponent
You may want to know whether a baseComponent instance (or one from
an inherited class) matches a given tag, title or description.
Matching a title/ description is to be intended as checking whether
the title / description has matching sub-strings.
A todoList might also be checked for its imminence, state, or
priority.

Now, the match(variable, value) method has been added to the
baseComponent class (and inherited classes) to check whether the
'variable' property matches the 'value' argument. 'variable' might
be, eg, 'title', 'description', 'tags', or, for the todoComponent,
'imminence', 'state', or 'priority'.

A proper matching function is selected based on 'value' as defined
in the filterCallbacks static object (which is extended in the
todoComponent with the 'imminence', 'state', or 'priority' fields).