Add getAllOfTypeNested(type) method to rootComponent, baseComponent, baseListComponent, and listInComponentMixin classes
This method retrieves all the instances of a given type ('T':
todoComponent, 'P': projectComponent, 'N': noteComponent) in the
whole descendants tree (eg, inboxProject project and
customProjectsList list for the rootComponent instances).

Note that here you can exploit the fact that all elements in a
todoListComponent, projectListComponent, noteListComponent have
the same type, so you don't need to test the condition for each
element in there. To quickly get the type of elements in the list,
a 'type' property has been added to such lists.

Also, for the baseComponent class, two methods are defined:
getAllOfTypeNested(type) and getAllOfTypeNested_withoutThis(type).
The former calls the latter, and perform the test for the current
instance. The latter is extended through the listInComponentMixin
and doesn't check the current instance, so it is used in the
recursive calls, where the instance itself is already considered,
as it is part of a list.

Again, some temporary test code has been added to initWebpage.js,
which will be removed later on.