Add editable property to baseComponent
This property is an indicator whether the baseComponent (or
inherited) instance should be editable. Note that this property
does not disable methods to modify the object. Instead, it might
be used to not build the UI to do that.

Indeed, this is what is done in the baseDomComponent class.

In general, all instances are editable. The only exception is the
Inbox project, which the user cannot edit.