Add editable property to baseListComponent
This property is an indicator whether the baseListComponent (or
inherited) instance should be editable. Note that this property
does not disable methods to modify the object. Instead, it might
be used to not build the UI to do that.

Indeed, this is what is done in the baseListDomComponent class.

In general, all list instances are editable. The only exception
will be the instances to show the results of the filtering or
search, which the user cannot edit.