Add listsToExclude parameter to initAllLists method defined in listInComponentMixin module
The listsToExclude parameter can be used to exclude the creation
of lists (among the default ones).

Recall that initAllLists() method is added to projectComponent,
todoComponent, noteComponent classes through composition. The
listsToExclude parameter is passed to the initAllLists method
through their constructor.

In general, all default lists are created for each component. The
only exception is the Inbox project, which has no sub-projects
that can be created.