Set icon to baseComponent and show it in the baseDomComponent
An icon can be set as a property of a baseComponent instance (or
inherited class instance). You can set it through the constructor.
Otherwise, a default one is set, based on the static property
'icon' of the class. Note that by default this is null, except for
the projectComponent, where a custom icon is set.

A getter icon() method is defined to get the right icon.

In the baseDomComponent, the icon is shown next to the title (note
that no icon is shown if this is null).