Add nTodoNested property in baseComponent class
This is just a counter of the number of not completed todo in the
descendants of the instance of baseComponent (or inherited) class.

It allows to quickly access this number without having to perform
a search in the whole descendant tree. This information can be
used to inform the user about the number of tasks not completed
yet, which is the core of the website.

There are methods to increase / decrease this number, which
recursively update the nTodoNested of all its ancestors.

The increasing/decreasing is to be performed when a todo instance
is created or deleted, based on its state, as well as when its
state changes.