Add nTodoNested property in baseListComponent class
This is just a counter of the number of not completed todo in the
descendants of all the items in baseListComponent (or inherited
class) instance. It allows to quickly access this number without
having to perform a search in the whole descendant tree. This
information can be used to inform the user about the number of
tasks not completed yet, which is the core of the website.

There are methods to increase / decrease this number.

The increasing/decreasing is to be performed when the
increaseNTodoNested() and decreaseNTodoNested() methods are called
on its item.

Also, when deleting an item of the list, you have to decrease both
the nTodoNested property of the current list and its parent by an
amount of item.nTodoNested.