Add generiBaseDomComponent module
This module instantiates dom instances that are even more generic
than baseDomComponent instances. Indeed, they are composed by just
a title, an icon, the path and the close button.

The baseDomComponent class has been refactored to inherit from
this generiBaseDomComponent class.

The generiBaseDomComponent class will be used to define a class
for generating dom views of the results of filtering todo / notes /
projects.