Add insertItem() method to baseListComponent class
The insertItem() method as been added to the baseListComponent
class. This is different from the already-defined addItem() method
because it inserts an already-existing item to the list.

The addItem() method has been refactored to use this.
A refactoring has been done also in the todoListComponent class.

To reflect this change, the listInComponentMixin module now also
creates the following methods: insertToProjectList(item),
insertToTodoList(item), insertToNoteList(item).

Moreover, it creates some generic verisons of some of the methods,
eg, insertToList(type,obj), where type is the type ('P', 'N', 'T')
of the list.

This insertItem() method will be used to create lists of filtered
data (which are already instantiated in the root component).