Add resultsComponent module
This module defines a class which is used to instantiate objects
holding the result of filtering / search within the root tree,
based on the provided parameters, and sort them in some
projectListComponent, todoListComponent, noteListComponent lists.

This class inherits from the genericBaseComponent class.

Some temporary code, which is currently used just for testing and
will be removed later on, has been added to the homeDomComponent
class.