Add resultsDomComponent module
This module defines a class which is used to instatiate a dom view
of the results of a filterBy/search operation.

It represent the object results (sorted as in resultsComponent) by
baseListDomComponent (or inherited class) list dom instances, in
which it is not possible to add new items.

Note that currently the syncronization between elements does not
work. Eg, deleting an todo in a todoDomComponent view does not
automatically update the view of the filtering results, if this is
currently displayed, too, and includes the todo to remove.

Some temporary test code has been added to homeDomComponent class,
which will be removed later on.