Fix items removal from baseListDomComponent displayed lists
Since the introduction of filterBy/search results views, an item
could be present in more than one displayed list. Before this
commit, removing such item only removed the item from one list.
Now, all the lists containing it are updated.

The removal handling is still handled via PubSub. However, a
different approach is used.

First of all, in each dom list there is a #ListMap Map object which
stores (obj, div) pairs, where obj is a reference to a
baseComponent (or inherited class) instance and used as the key,
and div the div element representing it.

A [list].getPubSubName("REMOVE ITEM", "main") PubSub token (which
contains the type/id of the parent, and the list type) is
published whenever an item is removed from a baseListComponent
(or inherited class) instance. Each baseListDomComponent (or
inherited class) instance is subscribed to the corresponding token
of this type, and removes the item from the dom when the right
token is emitted.

A different [list].getPubSubName("REMOVE ITEM", "main", false)
PubSub token (which only contains the type of the list) has been
used to notify other lists which contain the item to be deleted.
Ie, those other lists are the ones in the search / filterBy
results components.

Those tokens are emitted when an item is removed from a
baseListComponent (or inherited class) instance, and one is
emitted for each descendant as well (this is handled via
recursion).

This whole emission of PubSub is done by default, but disabled
based on a parameter 'notify' in the removeItem() method, which
is set to false when items are removed from lists in
resultsComponent instances (to avoid sending duplicated tokens).

Then, any list of the same type subscribed to this token checks
whether obj is being represented in there (by searching in the
.#ListMap) and removes it if present.